// Code/UICode/src/app/Interfaces/statusObject.ts
export class Status {
    ObjectType: String;
    ObjectId: String;
    BeaconId: String;
    BeaconStatus: String;
    BeaconTimestamp: String;
    Temperature: Number;
    Humidity: Number;
    TemperatureAlert: boolean;
    HumidityAlert: boolean;
    TamperAlert: boolean;
    ShockVibrationAlert: boolean;

}